moved more xem commands from main to new server
