+++ tests/test_server/test_server_routes_populate_queues.py
@@ -277,3 +277,126 @@ def test_send_single_start_acquisition_command__populates_queue(
     response_json = response.get_json()
     assert response_json["command"] == "start_acquisition"
     assert response_json["suppress_error"] is True
+
+
+def test_send_single_get_serial_number_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    response = test_client.get("/insert_xem_command_into_queue/get_serial_number")
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "get_serial_number"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "get_serial_number"
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_get_device_id_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    response = test_client.get("/insert_xem_command_into_queue/get_device_id")
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "get_device_id"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "get_device_id"
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_is_spi_running_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    response = test_client.get("/insert_xem_command_into_queue/is_spi_running")
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "is_spi_running"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "is_spi_running"
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_read_from_fifo_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    # test_bytearray = produce_data(1, 0)
+    # fifo = Queue()
+    # fifo.put(test_bytearray)
+    # queues = {"pipe_outs": {PIPE_OUT_FIFO: fifo}}
+    # simulator = FrontPanelSimulator(queues)
+    # simulator.initialize_board()
+    # simulator.start_acquisition()
+    # ok_process = test_process_manager.get_ok_comm_process()
+    # ok_process.set_board_connection(0, simulator)
+
+    response = test_client.get(
+        "/insert_xem_command_into_queue/read_from_fifo?num_words_to_log=72"
+    )
+    assert response.status_code == 200
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "read_from_fifo"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "read_from_fifo"
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_read_from_fifo_command_with_hex_notation__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    # test_bytearray = produce_data(1, 0)
+    # fifo = Queue()
+    # fifo.put(test_bytearray)
+    # queues = {"pipe_outs": {PIPE_OUT_FIFO: fifo}}
+    # simulator = FrontPanelSimulator(queues)
+    # simulator.initialize_board()
+    # simulator.start_acquisition()
+    # ok_process = test_process_manager.get_ok_comm_process()
+    # ok_process.set_board_connection(0, simulator)
+
+    response = test_client.get(
+        "/insert_xem_command_into_queue/read_from_fifo?num_words_to_log=0x48"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "read_from_fifo"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "read_from_fifo"
+    assert response_json["suppress_error"] is True
